app/database/requirements.txt
python-dotenv==1.1.0
fastapi==0.115.9
jinja2==3.1.6
psycopg2_binary==2.9.10
sqlalchemy==2.0.41
uvicorn==0.35.0
supabase==2.15.1